feat: generate docs for the other commands as well
